requirements.txt
django==5.2.4
djangorestframework==3.16.0
drf-spectacular==0.28.0
django-extensions==4.1
graphviz==0.21
pydotplus==2.0.2
djangorestframework-simplejwt==5.5.1